add multiple Tasks Features

+++ ShellBackup.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 CONFIG_FILE="$(dirname "$0")"/ShellBackup.conf
+TASKS="$(dirname "$0")"/Tasks.conf
+DATE=$(date +%Y-%m-%d)
+NOW=$(date +%Y%m%d_%H%M)
 
 if [ -f $CONFIG_FILE ]; then
     source $CONFIG_FILE
@@ -19,15 +22,41 @@ else
     fi
 fi
 
-DATE=$(date +%Y-%m-%d)
-BACKUP_PATH="$PARENT_DIR/changed/$DATE"
-NOW=$(date +%Y%m%d_%H%M)
+if [ -f $TASKS ]; then
+    while IFS="|" read -r NAME PARENT_DIR TARGET_DIR SOURCE_DIR; do
+        if [[ $NAME == *\#*  ]]; then
+            echo "Skipping comment: $NAME"
+            continue
+        fi
+        if [[ -z "$PARENT_DIR" || -z "$TARGET_DIR" ]]; then
+            echo "Skip Non-Tasks: $NAME"
+            continue
+        fi
+        echo "Handleing $NAME..."
+        # mkdir -p $TARGET_DIR
+        SOURCE_DIR=$(eval echo $SOURCE_DIR)
+        TARGET_DIR=$(eval echo $TARGET_DIR)
+        PARENT_DIR=$(eval echo $PARENT_DIR)
+        TARGET_DIR="$PARENT_DIR/$TARGET_DIR"
+        BACKUP_PATH="$PARENT_DIR/changed/$DATE"
+        echo $PARENT_DIR
+        echo $SOURCE_DIR
+        echo $TARGET_DIR
+        echo $BACKUP_PATH
+        rsync -av --delete \
+              --backup \
+              --backup-dir="$BACKUP_PATH" \
+              --suffix="_$NOW" \
+              --exclude '.git' \
+              "$SOURCE_DIR" "$TARGET_DIR"
+    done < $TASKS
+else
+    echo "Error: Can't find Tasks file $TASKS"
+    exit 1
+fi
 
-mkdir -p $TARGET_DIR
 
-rsync -av --delete \
-      --backup \
-      --backup-dir="$BACKUP_PATH" \
-      --suffix="_$NOW" \
-      "$SOURCE_DIR" "$TARGET_DIR"
+
+
+
